Refactor auth service structure in the medication manager service gateway
