+++ apps/medication-manager/src/users/users.controller.ts
@@ -1,21 +1,9 @@
-import { Body, Controller, Get, Param, Post } from '@nestjs/common';
-import { UsersService } from './users.service';
-import { CreateUserDto } from './dto/create-user.dto';
-import { VerifyEmailDto } from '../../../users/src/dto/verify-email.dto';
+import { Controller, Get, Param } from '@nestjs/common';
+import { UsersService } from './users.service';;
 
 @Controller('users')
 export class UsersController {
   constructor(private readonly usersService: UsersService) {}
-
-  @Post('register')
-  async register(@Body() createUserDto: CreateUserDto) {
-    return await this.usersService.register(createUserDto);
-  }
-
-  @Post('verify-email')
-  async verifyEmail(@Body() verifyEmailDto: VerifyEmailDto) {
-    return await this.usersService.verifyEmail(verifyEmailDto);
-  }
   @Get('/:id')
   async findUserById(@Param('id') id: string) {
     return await this.usersService.findUserById(id);
