+++ apps/medication-manager/src/users/users.service.ts
@@ -15,65 +15,6 @@ import { VerifyEmailDto } from '../../../users/src/dto/verify-email.dto';
 export class UsersService {
   private readonly logger = new Logger(UsersService.name);
   constructor(@Inject(USERS_SERVICE) private readonly client: ClientProxy) {}
-  async register(createUserDto: CreateUserDto) {
-    try {
-      const response = await firstValueFrom(
-        this.client.send({ cmd: 'register_user' }, createUserDto).pipe(
-          timeout(30000),
-          catchError((err) => {
-            this.logger.error(
-              `Microservice communication error: ${err.message}`,
-            );
-            throw new InternalServerErrorException(
-              'Failed to communicate with user service. Please try again later.',
-            );
-          }),
-        ),
-      );
-      if (response?.error) {
-        throw new BadRequestException(response.error);
-      }
-
-      return response;
-    } catch (error) {
-      if (error instanceof BadRequestException) {
-        throw error;
-      }
-      throw new InternalServerErrorException(
-        'An error occurred while registering the user. Please try again later.',
-      );
-    }
-  }
-
-  async verifyEmail(verifyEmailDto: VerifyEmailDto) {
-    try {
-      const response = await firstValueFrom(
-        this.client.send({ cmd: 'verify_email' }, verifyEmailDto).pipe(
-          timeout(30000),
-          catchError((err) => {
-            this.logger.error(
-              `Microservice communication error: ${err.message}`,
-            );
-            throw new InternalServerErrorException(
-              'Failed to communicate with user service. Please try again later.',
-            );
-          }),
-        ),
-      );
-      if (response?.error) {
-        throw new BadRequestException(response.error);
-      }
-
-      return response;
-    } catch (error) {
-      if (error instanceof BadRequestException) {
-        throw error;
-      }
-      throw new InternalServerErrorException(
-        'An error occurred while verifying email. Please try again later.',
-      );
-    }
-  }
   async findUserById(id: string) {
     try {
       const response = await firstValueFrom(
